03/12

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: mbellini <mbellini@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/23 14:41:31 by mbellini          #+#    #+#             */
-/*   Updated: 2021/12/03 10:32:04 by mbellini         ###   ########.fr       */
+/*   Updated: 2021/12/03 16:42:55 by mbellini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	if (!dest && !src)
 		return (NULL);
 	if (dest > src)
 	{
-		i = (int)n - 1;
+		i = n - 1;
 		while (i >= 0)
 		{
 			*(char *)(dest + i) = *(char *)(src + i);
@@ -30,7 +30,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 	else
 	{
 		i = 0;
-		while (i < (int)n)
+		while (i < n)
 		{
 			*(char *)(dest + i) = *(char *)(src + i);
 			i++;
